Move series’ number from comic to record (#63).
