+++ Model/src/main/sql/postgresql-3.1-migration.sql
@@ -1,0 +1,6 @@
+-- Migration from a 3.0 schema to a 3.1 schema
+
+-- Move `number` column from comics to records.
+alter table records add column number int;
+update records r set number = (select c.number from comics c where c.id = r.id);
+alter table comics drop column number;
